+++ backend/framework/domain/src/main/java/io/metersphere/api/domain/ApiDebugModuleExample.java
@@ -464,52 +464,52 @@ public class ApiDebugModuleExample {
             return (Criteria) this;
         }
 
-        public Criteria andPosEqualTo(Integer value) {
+        public Criteria andPosEqualTo(Long value) {
             addCriterion("pos =", value, "pos");
             return (Criteria) this;
         }
 
-        public Criteria andPosNotEqualTo(Integer value) {
+        public Criteria andPosNotEqualTo(Long value) {
             addCriterion("pos <>", value, "pos");
             return (Criteria) this;
         }
 
-        public Criteria andPosGreaterThan(Integer value) {
+        public Criteria andPosGreaterThan(Long value) {
             addCriterion("pos >", value, "pos");
             return (Criteria) this;
         }
 
-        public Criteria andPosGreaterThanOrEqualTo(Integer value) {
+        public Criteria andPosGreaterThanOrEqualTo(Long value) {
             addCriterion("pos >=", value, "pos");
             return (Criteria) this;
         }
 
-        public Criteria andPosLessThan(Integer value) {
+        public Criteria andPosLessThan(Long value) {
             addCriterion("pos <", value, "pos");
             return (Criteria) this;
         }
 
-        public Criteria andPosLessThanOrEqualTo(Integer value) {
+        public Criteria andPosLessThanOrEqualTo(Long value) {
             addCriterion("pos <=", value, "pos");
             return (Criteria) this;
         }
 
-        public Criteria andPosIn(List<Integer> values) {
+        public Criteria andPosIn(List<Long> values) {
             addCriterion("pos in", values, "pos");
             return (Criteria) this;
         }
 
-        public Criteria andPosNotIn(List<Integer> values) {
+        public Criteria andPosNotIn(List<Long> values) {
             addCriterion("pos not in", values, "pos");
             return (Criteria) this;
         }
 
-        public Criteria andPosBetween(Integer value1, Integer value2) {
+        public Criteria andPosBetween(Long value1, Long value2) {
             addCriterion("pos between", value1, value2, "pos");
             return (Criteria) this;
         }
 
-        public Criteria andPosNotBetween(Integer value1, Integer value2) {
+        public Criteria andPosNotBetween(Long value1, Long value2) {
             addCriterion("pos not between", value1, value2, "pos");
             return (Criteria) this;
         }
